requirements.txt
Django==1.10
django-extensions==1.6.7
djangorestframework==3.4.3
Pillow==3.3.0
pip==8.1.2
psycopg2==2.6.2
pytz==2016.6.1
setuptools==25.1.6
six==1.10.0
wheel==0.29.0
